+++ CHANGELOG.md
@@ -97,7 +97,7 @@
 - Dev: Fixed incorrect lua generation of static methods for typescript plugins. (#6190, #6223)
 - Dev: Merged top/bottom and left/right notebook layouts. (#6215)
 - Dev: Refactored `Button` and friends. (#6102, #6255, #6266, #6302, #6268, #6334, #6371, #6372)
-- Dev: Made "add split" button (part of the split header) a natively rendered button. (#6349)
+- Dev: Made "add split" button (part of the split header) a natively rendered button. (#6349, #6385)
 - Dev: Made Settings & Account button on Linux/macOS SVGs. (#6267)
 - Dev: Made "Chatters" button an SVG. (#6397)
 - Dev: Made "Moderation" button an SVG. (#6398, #6412, #6414)